Workaround to prevent deadlocks for parallel runs in Python 3 (NOT WORKING)
Python 3 uses hashed dictionaries. The order in which elements of such
dictionaries are listed is not unique (on different processes) and this causes
problems in destructors, especially when destructing objects with parallel
semantic.

One possible workaround is to use 'OrderedDict' which remembers the order of
insertion of the elements.

Note that the problem doesn't have to occur due to explicitly defined
dictionaries. (Python uses dictionaries internally to store attributes.)