Tune parameters for 'pool' benchmark
Values of the mobility coefficient 'M0' were experimentally found as the lowest
values with which one can handle the computation for the test case 2 with
'SemiDecoupled' scheme (OTD = 1) on a coarse mesh (hmin = 1/32).

It seems that if we want to use constant mobility, it must be chosen two
orders smaller than in the degenerate case.

'FullyDecoupled' scheme has problems on the coarse mesh. It helps if we turn
off the parameter 'omega_2'.

Proper tests on a sufficiently fine mesh are needed!